Resolved type ambiguity using asTypeOf
'daily' is a type constructor and, as used, is ambiguous
in its type. Since the type construction happens first and
then record update occurs there is no way for GHC to know
that the data should have the same type 'a' as the field
being updated. This is fixed by constraining the type using
asTypeOf.

Other possibilities are to turn on ScopedTypeVariables, and
explicity type 'daily' inline. Or one could even use a concrete
version like 'dailyUTC' since the record update would re-adjust
the data to the correct concrete type used elsewhere in the body
of the code. No option seems must better than any other and
I chose the simples, and most portable.

Thanks to 'aavogt' and 'dmwit' on #haskell for the explanation.